Update stripe-deploy.md
Hi team, I noticed an error when deploying the script.

When copying the WebApp URL from the Azure Portal, the https:// part is not automatically copied. This leads to Stripe not recognizing it as a valid URL and thus the product subscriptions in the DevPortal are then not reflected in APIM nor in the DevPortal user profile.

Therefore, I suggest adding the https:// part here explicitly.

Stripe log error:
url_invalid - url
Invalid URL: An explicit scheme (such as https) must be provided.